# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Academic Bridge Challenge API",
        "description": "RESTful API for Employee Attendance Management System. This API provides endpoints for user authentication, employee management, attendance tracking, and report generation.",
        "contact": {
            "name": "API Support",
            "email": "support@academicbridge.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server (Versioned)"
        }
    ],
    "paths": {
...
        "/v1/reports/attendance/pdf": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate attendance report in PDF format",
                "operationId": "0ab0c018dcf3bf000528535c3ef2471b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Report date (Y-m-d), defaults to today",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Report start date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Report end date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file download",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/reports/attendance/excel": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate attendance report in Excel format",
                "operationId": "4ea0e7d55d9d5543700988da99a196e2",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Report date (Y-m-d), defaults to today",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Report start date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Report end date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}